backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
scikit-learn==1.4.2
numpy==1.26.4
pdfminer.six==20221105
python-docx==1.1.2
PyJWT==2.8.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
